# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SP MLM Platform API",
        "description": "API Documentation for Mobile App and Frontend",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Main Server"
        }
    ],
    "paths": {
...
        "/api/user/recharge/config": {
            "get": {
                "tags": [
                    "User Recharge"
                ],
                "summary": "Get UPI/USDT details for recharge",
                "operationId": "c09dd50d6fbf99e09c8278ff1e733cd7",
                "responses": {
                    "200": {
                        "description": "Configuration fetched successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/recharge": {
            "post": {
                "tags": [
                    "User Recharge"
                ],
                "summary": "Submit a new recharge request",
                "operationId": "9de80693c106b7570009f6f5a46e4533",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "method",
                                    "screenshot"
                                ],
                                "properties": {
                                    "amount": {
                                        "description": "Minimum amount usually 500",
                                        "type": "number",
                                        "example": 1000
                                    },
                                    "method": {
                                        "type": "string",
                                        "enum": [
                                            "upi",
                                            "usdt"
                                        ],
                                        "example": "upi"
                                    },
                                    "utr": {
                                        "description": "Required if method is upi (12 digit UTR number)",
                                        "type": "string",
                                        "example": "123456789012"
                                    },
                                    "hash": {
                                        "description": "Required if method is usdt (TxHash)",
                                        "type": "string",
                                        "example": "Txabcdef123456789"
                                    },
                                    "screenshot": {
                                        "description": "Payment screenshot image file",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Recharge request submitted"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/recharge/history": {
            "get": {
                "tags": [
                    "User Recharge"
                ],
                "summary": "Get user recharge history",
                "operationId": "ed1d63486bb7308d9c9e35d3ae0f9418",
                "responses": {
                    "200": {
                        "description": "History fetched successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/settings/change-password": {
            "post": {
                "tags": [
                    "User Settings"
                ],
                "summary": "Change Account Password",
                "operationId": "ad6aa43779f5401cf3390a96216a0166",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "new_password",
                                    "new_password_confirmation"
                                ],
                                "properties": {
                                    "current_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "oldsecret123"
                                    },
                                    "new_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newsecret123"
                                    },
                                    "new_password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newsecret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password updated successfully"
                    },
                    "400": {
                        "description": "Incorrect current password"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/settings/transactions": {
            "get": {
                "tags": [
                    "User Settings"
                ],
                "summary": "Get User Transactions (Passbook)",
                "operationId": "07d1ab7f389e6835fdf8b0d208ba63b8",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page (default: 20)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transactions fetched successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/tasks": {
            "get": {
                "tags": [
                    "User Tasks"
                ],
                "summary": "Get Daily Task Status",
                "operationId": "c2f95901d0cdf426ce7294eae59ac118",
                "responses": {
                    "200": {
                        "description": "Task data fetched successfully"
                    },
                    "400": {
                        "description": "No active level found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/tasks/submit": {
            "post": {
                "tags": [
                    "User Tasks"
                ],
                "summary": "Submit a task and earn",
                "operationId": "b3ec2d5daa32d7ade72c1f3cdfd81147",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rating"
                                ],
                                "properties": {
                                    "rating": {
                                        "description": "Star rating between 1 to 5",
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task completed successfully"
                    },
                    "400": {
                        "description": "Limit reached or error"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/team": {
            "get": {
                "tags": [
                    "User Team"
                ],
                "summary": "Get User's 3-Level Team Network",
                "operationId": "d2448f40e67f709b6689e306224c9c60",
                "responses": {
                    "200": {
                        "description": "Team network data fetched successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Login with Mobile and Password to get the authentication token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}